assistant/requirements.txt
fastapi
uvicorn
requests
pydantic
paramiko
psutil
chardet
python-magic
GPUtil
httpx
sse-starlette
